django-platform-auth/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
Pillow>=10.0
python-decouple>=3.8
PyJWT>=2.8
cryptography>=41.0
